Replace use of POSIX ssize_t data type
Commit UUID: ed5b40cd-ed1e-4be9-9136-e4348b7cca87

+++ trunk/include/matroska.h
@@ -24,7 +24,7 @@ typedef struct matroska_state *matroska_hdl_t;
 typedef struct {
     int (*open)(void **, int, void *);
     int (*close)(void *);
-    int (*read)(void *, void *, ssize_t *);
+    int (*read)(void *, void *, int64_t *);
     int (*write)(void *, const void *, size_t);
     int (*sync)(void *);
     int (*get_fpos)(void *, int64_t *);
